# api/openapi.yaml
openapi: 3.0.3
info:
  title: Packbase API
  description: API for managing a hiking gear inventory.
  version: 0.1.0

servers:
  - url: http://localhost:8080/api/v1
    description: Local development

tags:
  - name: Auth
    description: Registration, login and session handling
  - name: Gear
    description: Hiking gear inventory items (the user's library)
  - name: Lists
    description: Packing lists that reference library items

# Authentication is a server-side session in an HttpOnly cookie. Unsafe requests
# (POST/PUT/DELETE) must also send the CSRF token from the XSRF-TOKEN cookie in
# the X-XSRF-TOKEN header. Any request (e.g. GET /auth/me) sets that cookie.
security:
  - sessionCookie: []

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      description: Does not log the user in; call /auth/login afterwards.
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email is already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags: [Auth]
      summary: Log in and start a session
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in; the session cookie is set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags: [Auth]
      summary: End the current session
      operationId: logout
      responses:
        '204':
          description: Logged out (also when there was no session)

  /auth/me:
    get:
      tags: [Auth]
      summary: Get the logged-in user
      operationId: getCurrentUser
      responses:
        '200':
          description: The logged-in user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /gear:
    get:
      tags: [Gear]
      summary: List all gear items
      operationId: listGear
      responses:
        '200':
          description: A list of gear items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GearItem'
    post:
      tags: [Gear]
      summary: Add a new gear item
      operationId: createGear
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GearItemInput'
      responses:
        '201':
          description: Gear item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GearItem'
        '400':
          $ref: '#/components/responses/BadRequest'

  /gear/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Gear]
      summary: Get a single gear item
      operationId: getGearById
      responses:
        '200':
          description: The gear item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GearItem'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Gear]
      summary: Update a gear item
      operationId: updateGear
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GearItemInput'
      responses:
        '200':
          description: Gear item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GearItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Gear]
      summary: Delete a gear item
      operationId: deleteGear
      responses:
        '204':
          description: Gear item deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /lists:
    get:
      tags: [Lists]
      summary: List the user's packing lists, newest first
      operationId: listPackLists
      responses:
        '200':
          description: The packing lists, each with its ordered entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PackList'
    post:
      tags: [Lists]
      summary: Create a packing list
      operationId: createPackList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackListInput'
      responses:
        '201':
          description: List created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackList'
        '400':
          $ref: '#/components/responses/BadRequest'

  /lists/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Lists]
      summary: Get a packing list
      operationId: getPackListById
      responses:
        '200':
          description: The packing list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackList'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Lists]
      summary: Replace a packing list's name and entries
      description: >
        The entries array is the complete, ordered content of the list. Every
        itemId must be a library item of the current user and appear only once.
      operationId: updatePackList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackListInput'
      responses:
        '200':
          description: List updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Lists]
      summary: Delete a packing list (library items are kept)
      operationId: deletePackList
      responses:
        '204':
          description: List deleted
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          example: hiker@example.com
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 128

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          maxLength: 128

    UserProfile:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email

    GearItemInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Titanium spork
        category:
          type: string
          maxLength: 100
          default: ''
          description: Empty means uncategorized.
          example: Kitchen
        weightGrams:
          type: integer
          minimum: 0
          example: 12
        quantity:
          type: integer
          minimum: 1
          default: 1
        notes:
          type: string

    ListEntry:
      type: object
      required:
        - itemId
        - quantity
      properties:
        itemId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 9999
        category:
          type: string
          maxLength: 100
          description: >
            Per-list category override. Omitted means "use the item's category";
            an empty string means uncategorized.

    PackListInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Kungsleden, 5 days
        entries:
          type: array
          maxItems: 1000
          items:
            $ref: '#/components/schemas/ListEntry'

    PackList:
      type: object
      required:
        - id
        - name
        - createdAt
        - entries
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        entries:
          type: array
          items:
            $ref: '#/components/schemas/ListEntry'

    GearItem:
      allOf:
        - type: object
          required:
            - id
            - createdAt
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
        - $ref: '#/components/schemas/GearItemInput'

    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Gear item not found

  responses:
    Unauthorized:
      description: Not logged in
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'